refactor and move LinkIconButton to it's own component
