+++ frontend/packages/volto-light-theme/src/components/Blocks/Listing/SummaryTemplate.jsx
@@ -66,7 +66,7 @@ const SummaryTemplate = ({ items, linkTitle, linkHref, isEditMode }) => {
               })}
               key={item['@id']}
             >
-              <Card item={item} showLink={showLink}>
+              <Card item={showLink ? item : null}>
                 <ItemBodyTemplate item={item} />
               </Card>
             </li>
