+++ frontend/packages/volto-light-theme/src/components/Blocks/Teaser/DefaultBody.tsx
@@ -44,7 +44,7 @@ const TeaserDefaultTemplate = (props) => {
   };
 
   return (
-    <Card item={href} showLink={showLink} openLinkInNewTab={openLinkInNewTab}>
+    <Card item={showLink ? href : null} openLinkInNewTab={openLinkInNewTab}>
       <Card.Image
         src={url && !image?.image_field ? url : undefined}
         item={!data.overwrite ? href : { ...href, ...localOverrides }}
