+++ frontend/packages/volto-light-theme/src/primitives/Card/Card.tsx
@@ -1,14 +1,7 @@
 import * as React from 'react';
-import { useSelector } from 'react-redux';
-import type { GetSiteResponse } from '@plone/types';
 import ConditionalLink from '@plone/volto/components/manage/ConditionalLink/ConditionalLink';
 import cx from 'classnames';
 import type { ObjectBrowserItem } from '@plone/types';
-import linkSVG from '@plone/volto/icons/link.svg';
-import Icon from '@plone/volto/components/theme/Icon/Icon';
-import { Button } from '@plone/components';
-import { flattenToAppURL } from '@plone/volto/helpers/Url/Url';
-import { useLocation, useHistory } from 'react-router-dom';
 
 type BaseCardProps = {
   /** Optional additional CSS class names to apply to the card. */
@@ -16,22 +9,17 @@ type BaseCardProps = {
   openLinkInNewTab?: boolean;
   children?: React.ReactNode;
 };
-type FormState = {
-  site: { data: GetSiteResponse };
-};
 
 type CardPropsWithItem = BaseCardProps & {
   /** List of items rendered within the card. Mutually exclusive with `href`. */
   href?: never;
   item: Partial<ObjectBrowserItem>;
-  showLink?: boolean;
 };
 
 type CardPropsWithoutItem = BaseCardProps & {
   /** Optional URL to make the card clickable as a link. */
   href?: string | undefined | null;
   item?: never;
-  showLink?: boolean;
 };
 
 type CardProps = CardPropsWithItem | CardPropsWithoutItem;
@@ -56,51 +44,15 @@ const childrenWithProps = (children, extraProps) => {
     return child;
   });
 };
-const useLinkIconNavigation = (item?: Partial<ObjectBrowserItem>) => {
-  const location = useLocation();
-  const history = useHistory();
 
-  const handleLinkIconClick = (e: React.MouseEvent<HTMLButtonElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    const targetUrl = item?.['@id'];
-    if (targetUrl) {
-      const flattenedTargetUrl = flattenToAppURL(targetUrl);
-      const searchParams = new URLSearchParams();
-      searchParams.set('return_to', location.pathname);
-
-      history.push({
-        pathname: flattenedTargetUrl,
-        search: searchParams.toString(),
-      });
-    }
-  };
-  return handleLinkIconClick;
-};
-
-const LinkIconButton = ({ item }: { item?: Partial<ObjectBrowserItem> }) => {
-  const site = useSelector<FormState, GetSiteResponse>(
-    (state) => state.site?.data,
-  );
-  const hideProfileLinks = site?.['kitconcept.disable_profile_links'];
-  const isPersonProfile = item?.['@type'] === 'Person' && hideProfileLinks;
-  const handleLinkIconClick = useLinkIconNavigation(item);
-  return (
-    isPersonProfile && (
-      <div className="card-link-icon">
-        <Button aria-label="link" onClick={handleLinkIconClick}>
-          <Icon name={linkSVG} size="33px" />
-        </Button>
-      </div>
-    )
-  );
-};
 const Card = (props: CardProps) => {
-  const { className, openLinkInNewTab, href, item, showLink } = props;
+  const hasItem = !!props.item;
+  const item = hasItem ? props.item : undefined;
+  const href = !hasItem ? props.href : undefined;
+  const { className, openLinkInNewTab } = props;
 
   const a11yLabelId = React.useId();
-  const isInteractive = !!props.href || (!!props.item && showLink !== false);
+  const isInteractive = !!props.href || !!props.item;
 
   const LinkToItem = React.useCallback(
     ({ children }: { children: React.ReactNode }) => {
@@ -125,7 +77,6 @@ const Card = (props: CardProps) => {
         {childrenWithProps(props.children, {
           a11yLabelId,
           LinkToItem,
-          item,
         })}
       </div>
     </div>
@@ -147,12 +98,10 @@ type CardImageProps = {
 
 const CardImage = (props: CardImageProps) => {
   const { src, item, image, imageComponent, showPlaceholderImage } = props;
-
   const Image = imageComponent || DefaultImage;
 
   return (
     <div className="image-wrapper">
-      <LinkIconButton item={item} />
       {src ? (
         <Image src={src} alt="" loading="lazy" responsive={true} />
       ) : item || image ? (
@@ -178,19 +127,16 @@ const CardImage = (props: CardImageProps) => {
 type CardSummaryProps = {
   /** The ID of the element that labels the card. */
   a11yLabelId?: string;
-  item?: Partial<ObjectBrowserItem>;
   LinkToItem?: React.ElementType;
   children?: React.ReactNode;
 };
 
 const CardSummary = (props: CardSummaryProps) => {
-  const { children, a11yLabelId, item, LinkToItem } = props;
-
+  const { a11yLabelId, LinkToItem } = props;
   return (
     <div className="card-summary">
-      <LinkIconButton item={item} />
-      {childrenWithProps(children, {
-        a11yLabelId: a11yLabelId,
+      {childrenWithProps(props.children, {
+        a11yLabelId,
         LinkToItem,
       })}
     </div>
